% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence}
\usage{
check_convergence(file, folder_wd = getwd(), starts_txt = "0;",
  type_imputation = FALSE, Rcond_min = 1e-06)
}
\arguments{
\item{file}{(integer)}

\item{starts_txt}{(character)}

\item{type_imputation}{(logical)}

\item{folder}{(character)}
}
\value{
(logical)
}
\description{
Scans an .out file from an Mplus model and looks for signs of convergence.
}
