% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rcond_S.R
\name{check_rcond_S}
\alias{check_rcond_S}
\title{Function to check Rcond and update it}
\usage{
check_rcond_S(Plist)
}
\arguments{
\item{Plist}{(list)}
}
\value{
(list)
}
\description{
Function to check Rcond and update it
}
