% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobs.R
\name{cprobs}
\alias{cprobs}
\title{Calculate posterior class probabilities}
\usage{
cprobs(Y_i, pi_vec, mu_mat, S_array, list_mdpattern = NULL,
  log = FALSE)
}
\arguments{
\item{Y_i}{(matrix) of size NxJ of observed values}

\item{pi_vec}{(numeric vector) with K elements containing component probabilities}

\item{mu_mat}{(matrix) of size JxK, where J is number of items}

\item{S_array}{(array) of size JxJxK with within class covariance matrix.}

\item{list_mdpattern}{(list) Optional. Generated from get_mdpatterns function.}
}
\value{
(matrix) of size NxK containing the posterior class porbabilties
}
\description{
This function calculates the posterior class probabiltieis from a finite mixture of multivariate normal distributions.
}
\examples{
cprobs(Y_i, pi_vec, mu_mat, S_array)
}
