% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobs_lcr.R
\name{cprobs_lcr}
\alias{cprobs_lcr}
\title{Calculate posterior class probabilities given a latent class regression}
\usage{
cprobs_lcr(Y_i, X_i, pi_vec, beta_array, S_array, list_mdpattern = NULL,
  fast = FALSE)
}
\arguments{
\item{Y_i}{(matrix) of size NxJ of observed values}

\item{X_i}{(matrix) Nx(P+1) design matrix. First column all 1's}

\item{pi_vec}{(numeric vector) with K elements containing component probabilities}

\item{beta_array}{(array) of size Jx(P+1)xK, where J is number of items, P is the number of predictors}

\item{S_array}{(array) of size JxJxK with within class residual covariance matrix.}

\item{list_mdpattern}{(list) Optional. Generated from get_mdpatterns function.}

\item{fast}{(logical) If TRUE, the use Rfast::dmvnorm. Otherwise use mixtools::dmvnorm}
}
\value{
(matrix) of size NxK containing the posterior class porbabilties
}
\description{
This function calculates the posterior class probabiltieis from a finite mixture of multivariate normal distributions.
}
\examples{
cprobs_lcr(Y_i, X_i pi_vec, beta_array, S_array)
}
