% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_naiveMplus_inpfile.R
\name{create_naiveMplus_inpfile}
\alias{create_naiveMplus_inpfile}
\title{Stratified Amelia Imputation}
\usage{
create_naiveMplus_inpfile(z, out_get_FMM, dffolderfiles, temp_wd_p,
  savedata = FALSE, type_imputation = FALSE, starts_txt = "0;",
  symbol_txt = "*")
}
\arguments{
\item{z}{(integer) Condition number}

\item{out_get_FMM}{(list) Gaussian finite mixture model parameters in format exported by get_FMM_params}

\item{dffolderfiles}{(data.frame) with the files and folders of the saved data}

\item{temp_wd_p}{(character). Specifies processor-specific temporary directory}

\item{savedata}{(logical) Save the cprobs and estiamtes (savedata = TRUE) or not (savedata = FALSE). Defaults to false.}

\item{type_imputation}{(logical) Add a "TYPE = IMPUTATION;" command in the Data command of Mplus. Defaults to FALSE (i.e. not TYPE statement in DATA command)}

\item{starts_txt}{(character) Custom starting values in Mplus (e.g "20 10;" correpsonds to "STARTS = 20 10;"). Defaults to "0;"}

\item{symbol_txt}{("character") Either "*" or the "(at)" symbol. The latter for fixing parameters.}
}
\value{
Saves the Mplus input files for Naive LPA analysis in the the folders specified of dffolderfiles input. Returns a character vector with the the Mplus code.
}
\description{
This function obtain imputations using Amelia, stratified by observed group membership
}
\examples{
create_naiveMplus_inpfile(z, data_conditions, dffolderfiles, temp_wd_p)
}
