% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2results.R
\name{df2results}
\alias{df2results}
\title{df2results}
\usage{
df2results(df, pop_params, dtype = "", p = NA, rep = NA, z = NA,
  pm = NA, pva = NA)
}
\arguments{
\item{dtype}{Defaults to ""}

\item{p}{Defaults to NA}

\item{rep}{Defaults to NA}

\item{z}{Defaults to NA}

\item{pm}{Defaults to NA}

\item{pva}{Defaults to NA}
}
\description{
What does this function do?
}
\examples{
lavaan2resultsa(obj_lavaan, pop_params)
}
