% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprobs.R, R/dmixmvrnorm.R
\name{dmixmvrnorm}
\alias{dmixmvrnorm}
\title{Calculate posterior class probabilities}
\usage{
dmixmvrnorm(Y_i, pi_vec, mu_mat, S_array, log = FALSE)

dmixmvrnorm(Y_i, pi_vec, mu_mat, S_array, log = FALSE)
}
\arguments{
\item{Y_i}{(matrix) of size NxJ of observed values}

\item{pi_vec}{(numeric vector) with K elements containing component probabilities}

\item{mu_mat}{(matrix) of size JxK, where J is number of items}

\item{S_array}{(array) of size JxJxK with within class covariance matrix.}

\item{log}{(logical) Whether the log density should be return. Defaults to FALSE.}

\item{Y_i}{(matrix) of size NxJ of observed values}

\item{pi_vec}{(numeric vector) with K elements containing component probabilities}

\item{mu_mat}{(matrix) of size JxK, where J is number of items}

\item{S_array}{(array) of size JxJxK with within class covariance matrix.}
}
\value{
(matrix) of size NxK containing the posterior class porbabilties

(vector) of size N containing (log) density values
}
\description{
This function calculates the posterior class probabiltieis from a finite mixture of multivariate normal distributions.

This function calculates the (log) density from a finite mixture of multivariate normal distributions.
}
\examples{
cprobs(Y_i, pi_vec, mu_mat, S_array)
dmixmvrnorm(Y_i, pi_vec, mu_mat, S_array, log = TRUE)
}
