% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMs_LPA.R
\name{EMs_LPA}
\alias{EMs_LPA}
\title{Imputation of MAR LPA Model using EM with Sampling}
\usage{
EMs_LPA(obsdf, M, z, data_conditions, starts0 = 20, bayes_boot = TRUE,
  boot_2x = FALSE, tempdir_mplus = getwd(), rep = NULL,
  itermax = 2, multiplier = 64, cl = NULL)
}
\arguments{
\item{obsdf}{(data.frame) Observed data}

\item{M}{(integer) Number of imputations to construct}

\item{z}{(integer) Data condition number}

\item{data_conditions}{(data.frame) Simulation conditions pertaining to the complete data}

\item{starts0}{(integer) Defaults to 20. Number of intial starts for Mplus when fitting FMM models.}

\item{bayes_boot}{(logical) Defaults to TRUE. Whether or not to perform Bayesian Bootstrap.}

\item{boot_2x}{(logical) Defaults to FALSE. Whether or not to implement a double Bootstrap.}

\item{tempdir_mplus}{(character) Defaults to that returned by getwd(). Temporary directory to store .dat, .inp, and .out files. Deleted upon successful termination.}

\item{rep}{(integer) Defaults to NULL. Replication number for temprorary directory naming purpose. If one not provided, then a random identifier is constructed.}

\item{itermax}{(integer) Defaults to 2. Number of iterations}

\item{multiplier}{(integer) Defaults to 64. Number to multiply starts by at each iteration.}

\item{cl}{Optional SOCKcluster object.}
}
\value{
(mids) An imputed data set of time (mids)
}
\description{
Imputation of MAR LPA Model using EM with Sampling
}
\examples{
EM_with_sampling(obsdf, M, z, data_conditions)
}
