% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMs_LPA_2_completion.R
\name{EMs_LPA_2_completion}
\alias{EMs_LPA_2_completion}
\title{Imputation by EM with Sampling}
\usage{
EMs_LPA_2_completion(obsdf, M, z, data_conditions, tmax = 30, ...)
}
\arguments{
\item{obsdf}{(data.frame) with the observed data}

\item{M}{(integer) Number of imputations}

\item{z}{(integer)}

\item{data_conditions}{(data.frame)}

\item{tmax}{(numeric). Defaults to a maximum of 30 minutes to achieve the imputations.}

\item{...}{Optional arguments to EMs_LPA() function (e.g., starts0, bayes_boot, boot_2x, tempdir_mplus, rep, itermax, multiplier, cl)}
}
\value{
A mids object with the imputations
}
\description{
Imputation by EM with Sampling
}
\examples{
EMs_LPA_2_completion<-function(obsdf, M, z, data_conditions)

}
