% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_class_proportions.R
\name{extract_class_proportions}
\alias{extract_class_proportions}
\title{Extract class proportions from Bayesian fitted LC model in Mplus .out file.}
\usage{
extract_class_proportions(file, path = getwd())
}
\arguments{
\item{file}{(character) The filename of the Mplus .out file where the starting values are located}

\item{path}{(character) The path where the file is located}
}
\value{
(character) vector with Model sytax at starting values from the .out file.
}
\description{
Extract class proportions from Mplus .out file.
}
\examples{
extract_class_proportions(file = "example.out")
}
