% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ICs_averaged.R
\name{extract_ICs_averaged}
\alias{extract_ICs_averaged}
\title{Get individual ICs for average type pooling method}
\usage{
extract_ICs_averaged(p, temp_wd_p_vec, kk, z, rep, starts_txt, pm = NA,
  pva = NA)
}
\arguments{
\item{p}{(integer) Processor number}

\item{kk}{(integer) Fitted number of classes}

\item{pm}{Defaults to NA}

\item{pva}{Defaults to NA}
}
\value{
(data.frame) results_df
}
\description{
Get individual ICs for average type pooling method
}
\examples{



}
