% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ICs_naive.R
\name{extract_ICs_naive}
\alias{extract_ICs_naive}
\title{Get ICs}
\usage{
extract_ICs_naive(TYPE = "None", DATATYPE = NULL, p, temp_wd_p_vec, kk,
  z, rep, starts_txt, pm = NA, pva = NA, type_imputation = FALSE)
}
\arguments{
\item{TYPE}{(character) Pooling type. Defaults to "NONE"}

\item{DATATYPE}{(charcter) One of (a) "Complete data", (b) "Observed data", (c) "Imputed data", or (d) "Stacked data"}

\item{kk}{(integer) Fitted number of classes}

\item{pm}{Defaults to NA}

\item{pva}{Defaults to NA}
}
\value{
(data.frame) results_df
}
\description{
Get ICs
}
\examples{



}
