% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ICs_stacked.R
\name{extract_ICs_stacked}
\alias{extract_ICs_stacked}
\title{Get individual ICs for stacked type pooling method}
\usage{
extract_ICs_stacked(p, temp_wd_p_vec, kk, z, rep, starts_txt, mids_pm_pva,
  pm = NA, pva = NA)
}
\arguments{
\item{p}{(integer) Processor number}

\item{kk}{(integer) Fitted number of classes}

\item{mids_pm_pva}{A mids object frmo get_imputed_ function}

\item{pm}{Defaults to NA}

\item{pva}{Defaults to NA}
}
\value{
(data.frame) results_df
}
\description{
Get individual ICs for stacked type pooling method
}
\examples{



}
