% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_svals.R
\name{extract_svals}
\alias{extract_svals}
\title{Extract starting values from Mplus .out file.}
\usage{
extract_svals(file, path = getwd())
}
\arguments{
\item{file}{(character) The filename of the Mplus .out file where the starting values are located}

\item{path}{(character) The path where the file is located}
}
\value{
(character) vector with Model sytax at starting values from the .out file.
}
\description{
Extract starting values from Mplus .out file.
}
\examples{
extract_svalues(file = "example.out")
}
