% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_do_rubinsrules.R
\name{fit_and_do_rubinrules}
\alias{fit_and_do_rubinrules}
\title{Fit mixture model to multiply imputed data and pool using Rubin's Rules (1987)}
\usage{
fit_and_do_rubinrules(dff_imputed, methods_list, pop_params_z,
  data_conditions, temp_wd_p_vec, p, z, pm, pva)
}
\arguments{
\item{dff_imputed}{(data.frame)}

\item{methods_list}{(list)}

\item{pop_params_z}{(list)}

\item{data_conditions}{(data.frame)}

\item{temp_wd_p_vec}{(vector) of characters}

\item{p}{(integer)}

\item{z}{(integer)}

\item{pm}{(integer)}

\item{pva}{(integer)}
}
\value{
(list) out_list with (a) parameters, (b) ARIV, (c) Vtot, (B)
}
\description{
Fit mixture model to multiply imputed data and pool using Rubin's Rules (1987)
}
\examples{
fit_and_do_rubinrules(dff_imputed, methods_list, pop_params_z, data_conditions, temp_wd_p_vec, p, z, pm, pva)

}
