% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_til_convergence.R
\name{fit_til_convergence}
\alias{fit_til_convergence}
\title{Fit mixture models in Mplus until convergence}
\usage{
fit_til_convergence(p, z, temp_wd_p_vec, target_wd, dff_target,
  pop_params_kk, starts0 = 20, mX_max = 7, type_imputation = FALSE,
  m = NA, ...)
}
\arguments{
\item{p}{Processor}

\item{z}{Data conditioning number}

\item{temp_wd_p_vec}{Processor-specific folder}

\item{target_wd}{Character of where data is located}

\item{dff_target}{List of files from the get_*_data functions}

\item{pop_params_kk}{Population parameters}

\item{starts0}{Must be a multiple of 10. Defaults to 20.}

\item{mX_max}{Highest exponents to go to (ie.e max mX for 2^mx)}

\item{type_imputation}{(logical) passed to create_NaiveMplus_inpfile() or check_convergence()}

\item{...}{Other arguments to send to create_naiveMplus_inpfile()}
}
\value{
(data.frame) results_df
}
\description{
Fit mixture models in Mplus until convergence
}
\examples{
fit_til_convergence<(z, target_wd, dff_target, pop_params_kk)
}
