% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma2pi.R
\name{gamma2pi}
\alias{gamma2pi}
\title{Convert conditional log-odds to probabilities}
\usage{
gamma2pi(gamma_vec)
}
\arguments{
\item{gamma_vec}{(numeric vector) of size K-1 with conditional log odds}
}
\value{
(numberic vector) of size K with corresponsding probabilites
}
\description{
Convert conditional log-odds to probabilities
}
\examples{
gamma2pi(c(0,0))
}
