% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tracker_and_outdirs.r
\name{gen_tracker_and_outdirs}
\alias{gen_tracker_and_outdirs}
\title{Generate Directories for Mplus .out Files and File Tracker}
\usage{
gen_tracker_and_outdirs(outwd, datwd, rdatawd, data_type, rep_vec, z_vec,
  data_conditions, fix_vec = FALSE, pm_vec = NA, pva_vec = NA,
  M = NA, dk_vec = 0, starts0 = 32, mkdir_out = TRUE)
}
\arguments{
\item{outwd}{(character) where to store the out files}

\item{datwd}{(character) where the dat files are located}

\item{rdatawd}{(character) where the Rdata file with the list resides}

\item{data_type}{(character) any of 'Complete data', 'Observed data', 'Stacked data' or 'Imputed data'}

\item{rep_vec}{(numeric) replications to track}

\item{z_vec}{(numeric) data conditions to track}

\item{data_conditions}{(data.frame) must include the variable z}

\item{fix_vec}{(logical) Either FALSE (free estimation) or 1 (fixed esitmation) or c(0,1) (i.e, both).}

\item{pm_vec}{(numeric) Defaults to NA. Percent missing identifier.}

\item{pva_vec}{(numeric) Defaults to NA. Imputation algorithm identfier}

\item{M}{(integer) Defaults to NA. Number of imputations.}

\item{dk_vec}{(numeric). Defaults to 0.}

\item{starts0}{(integer). Defaults to 32.}

\item{mkdir_out}{(logical). Defaults to TRUE. Whether or not to make the folder for where the outfiles will be stored.}
}
\value{
tracker_df (data.frame)
}
\description{
Generate Directories for Mplus .out Files and File Tracker
}
\examples{
get_tracker_and_outdir(outwd, datwd, data_type, rep_vec, z_vec, data_conditions)
}
