% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditional.R
\name{get_conditional}
\alias{get_conditional}
\title{Conditional distribution of lower-dimensional MVN from higher-dimensional MVN}
\usage{
get_conditional(mu, S, U, V)
}
\arguments{
\item{mu}{(numeric) Mean vector of full joint MVN distribution}

\item{S}{(matrix) Covariance matrix of full joint MVN distribution}

\item{U}{(numeric) Indices of missing variables}

\item{V}{(numeric) Indices of observed variables}
}
\value{
A list with the following elements multivariate regression covefficients (beta) and the residual variance ovariance matrics S_eps
}
\description{
Conditional distribution of lower-dimensional MVN from higher-dimensional MVN
}
\examples{
get_conditional(mu,S,U,V)
}
