% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diffICs.R
\name{get_diffICs}
\alias{get_diffICs}
\title{Get difference in ICs from enumeration summaries}
\usage{
get_diffICs(result_df, token = "")
}
\arguments{
\item{result_df}{(data.frame) giving summary results during enumeration with variables: (a) Parameters, (b)AIC, and (c) BIC}
}
\value{
(data.frame) with difference ICs appended to results_df input file
}
\description{
Obtain the population parameters given the data
}
\examples{
get_diffICs(LL_df)
}
