% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KL_DMAT.R
\name{get_KL_DMAT}
\alias{get_KL_DMAT}
\title{Obtain K-L Divergence discrepancy matrix across classes using population values}
\usage{
get_KL_DMAT(params, z, data_conditions)
}
\arguments{
\item{params}{() Parameter estimates obtained from Mplus}

\item{z}{(integer)}

\item{data_conditions}{(data.frame)}
}
\value{
(matrix) of size K-by-K with the KL divergences
}
\description{
Obtain K-L Divergence discrepancy matrix across classes using population values
}
\examples{
param_complete = get_KL_DMAT(params, z, data_conditions)
}
