% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdpatterns.R
\name{get_mdpatterns}
\alias{get_mdpatterns}
\title{Missing data patterns}
\usage{
get_mdpatterns(df)
}
\arguments{
\item{df}{(data.frame) of size N-by-J with missing values}
}
\value{
List with the following elements 
         (a) pattern_i - (data.frame) giving the missing data binary pattern (e.g., Rpatt) and numeric pattern identifier
         (b) table_r - (data.frame) providing missing data patterns present in the provided data frame
         (c) U_r - (list) A list of size equal to the number of patterns in table_r (minus 1 due to excluding complete data). Each 
                         elements contains the indices corresponding to missing values for a part
         (d) V_r = (list) Similar to U_r, except with observed data indices
}
\description{
Missing data patterns
}
