% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Plist.R
\name{get_Plist}
\alias{get_Plist}
\title{Format population parameters to P list}
\usage{
get_Plist(z, data_conditions)
}
\arguments{
\item{z}{(integer) condition identifier (number) for the complete data (as specified by data_conditions)}

\item{data_conditions}{(data.frame) simulation conditions pertaining to the complete data}
}
\value{
(list) with pi, mu, and S as arguments defining the mixture model
}
\description{
Format population parameters to P list
}
\examples{
get_Plist(z, data_conditions)
}
