% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psuedoc_draw.R
\name{get_pseudoc_draw}
\alias{get_pseudoc_draw}
\title{Psuedo-class draw LC membership}
\usage{
get_pseudoc_draw(CPROBS_df)
}
\arguments{
\item{CPROBS_df}{(data.frame) N-by-K data frame with posterior class probabilities for each individual}
}
\value{
class_vec (vector) Vector with N psuedoclass draws for each individual
}
\description{
Psuedo-class draw LC membership
}
