% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rotation_matrix.R
\name{get_rotation_matrix}
\alias{get_rotation_matrix}
\title{Rotation matrix}
\usage{
get_rotation_matrix(t = pi/4, dim = 2)
}
\arguments{
\item{t}{- (numeric) Rotation angle in radians. Defaults to pi/4.}

\item{dim}{(integer) Rank of rotation matrix. Defaults to 2 for rotating in R2.}
}
\value{
(Matrix) A rotation matrix of size dim-by-dim.
}
\description{
Construct a rotation matrix that rotates data in R2 (or R3, eventually)
}
\examples{
A = get_rotation_matrix(t = pi/3, dim = 2)
}
