% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stacked_data.R
\name{get_stacked_data}
\alias{get_stacked_data}
\title{Simulate complete LPA data for simulation study}
\usage{
get_stacked_data(z, list_imputed, methods_list, pctmiss_vec, save_it = FALSE,
  rep = NA, p = NA, temp_wd_p_vec = NULL)
}
\arguments{
\item{z}{(integer) condition identifier for the complete data (as specified by data_conditions)}

\item{list_imputed}{(list) from calling get_imputed_data() function}

\item{methods_list}{(list) List w/ pva elements. Designates the imputation methods to try.}

\item{pctmiss_vec}{(vector) Vector w/ pm elements. Designates percent missing simulation conditions}

\item{save_it}{(logical) if TRUE, then it saves a .dat data set to load into Mplus; rep, p, & temp_wd_p_vec need to be specified.}

\item{rep}{(integer) Replication number (defaults to 1).}

\item{p}{(integer) Processor number (defaults to 1).}

\item{temp_wd_p_vec}{(character vector). Required if save_it==TRUE.  processor-specific temporary directory}
}
\value{
out_list  (list) with the following elements
      (A) stack_df - (data.frame) stacked imputed data set
      (b) dffolderfiles (data.frame) with the files and folders of the saved data
}
\description{
This function simulates complete data.
}
\examples{
get_stacked_data(z,data_conditions,methods_list,pct_miss_vec,save_it = TRUE)
}
