% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tech1.R
\name{get_tech1}
\alias{get_tech1}
\title{Get tech1 estimates identifiers}
\usage{
get_tech1(out_readModels)
}
\arguments{
\item{out_readModels}{Object from MplusAutomations readModels() function.}
}
\value{
tech1_df (data.frame) with tech1 identifier for Mplus parameters
}
\description{
Get tech1 estimates identifiers
}
\examples{
tech1_df = get_tech1(out_readModels)
}
