% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ix.R
\name{ix}
\alias{ix}
\title{Index from results list}
\usage{
ix(x, head, p = NULL, z = NULL, rep = NULL, pm = NULL,
  pva = NULL, sub = TRUE)
}
\arguments{
\item{x}{(character vector) with names from a results list}

\item{head}{(character) how x should starte (e.g. "list_complete")}

\item{p}{Optional (integer) with processor number}

\item{z}{Optional (integer) with data condition}

\item{rep}{Optional (integer) with replication}

\item{pm}{Optional (integer) with percent missing}

\item{pva}{Optional (integer) with imputation algorithm}

\item{sub}{Substitute periods if there are spaces the elements of x. Done for correct search. Defaults to TRUE.}
}
\value{
vector of integers with indices in x that match.
}
\description{
Index from results list
}
