% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_check_switching.R
\name{KL_check_switching}
\alias{KL_check_switching}
\title{Identify label switching occured using KL divergence}
\usage{
KL_check_switching(nn, Qlist_x, Plist_x, name = "glpk", t_max = 30,
  approximate = 0, round_cplex = 0, trace_cplex = 0)
}
\arguments{
\item{nn}{(integer) row in tracker_df}

\item{Qlist_x}{(list) List with sample population estimates for LPA model}

\item{Plist_x}{(list) List with population values for LPA model}

\item{name}{(character) name of discrete optimization solver}

\item{t_max}{(numeric) maximum time}

\item{approximate}{0 or 1. Defaults to 0 indicating not ot use an approximate solution.}

\item{round_cplex}{See bmatch}

\item{trace_cplex}{See bmatch}
}
\description{
Identify label switching occured using KL divergence
}
\examples{
KL_check_switching(mu_est_mat, S_est_array, z,)
}
