% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meng_Rubin_D3.R
\name{Meng_Rubin_D3}
\alias{Meng_Rubin_D3}
\title{Calculate the Meng and Rubin's (1992) D-statistic (denoted D3) for a specified replication, data condition, number of imputations}
\usage{
Meng_Rubin_D3(pooled_tracker_df, rep_D3, z_D3, Mmax_D3)
}
\arguments{
\item{pooled_tracker_df}{(data.frame) Tracker with both free and constrained (fixed) loglikelihood values}

\item{rep_D3}{(integer) Replication to calculate D3}

\item{z_D3}{(integer) Data condition number to calculate D3}

\item{Mmax_D3}{(integer) Number of imputations to use. Note that a D3 may be calculated using fewer imputed data sets if not all data sets converged.}
}
\value{
(data.frame) A data frame with K-1 rows the following columns:
rep - (integer) Replication
z - (integer) data condition
M_condition - (integer) Number of requested to us imputations to use.
H0 -  Reduced model number of components.
H1 - Full model number of components.
dQ - Difference in number of parameters between full and reduced model.
M_D_3 - Number of imputed data sets used to calculate D_3 and ARIV_2 statistics. May be less than M_condition (and Mmax_D3) specified if the model failed converge for some of the data sets.
ARIV_2 - Average relative increase in variance statistic calculated a proposed by Meng and Rubin (1992)
v_4 - The denominator degrees of freedom for the F reference distribution to which the sampling distriubtion of D_3 is supposed to follow (see Meng & Rubin, 1992)
D_3 - The Meng and Rubin (1992) D-statistic. Should be nonnegative.
}
\description{
Calculate the Meng and Rubin's (1992) D-statistic (denoted D3) for a specified replication, data condition, number of imputations
}
\examples{
Meng_Rubin_D3(pooled_tracker_df = lpa_pool_tracker_df2, rep_D3 = unique(lpa_pool_tracker_df2$rep),z_D3 = 10, Mmax_D3 = 100)
}
