% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mplus2Qlist.R
\name{Mplus2Qlist}
\alias{Mplus2Qlist}
\title{Format Mplus Parameters to Q list}
\usage{
Mplus2Qlist(params_df)
}
\arguments{
\item{params_df}{(data.frame) Unstandardized parameters extracted from MplusAutomation}
}
\value{
(list) with pi, mu, and S as arguments defining the mixture model
}
\description{
Format Mplus Parameters to Q list
}
\examples{
Mplus2Q(params_df)
}
