% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_withinclass_Mplus_syntax.R
\name{naive_withinclass_Mplus_syntax}
\alias{naive_withinclass_Mplus_syntax}
\title{Create the within}
\usage{
naive_withinclass_Mplus_syntax(mu_k, S_k, usev_vec,
  Model_Covariance = FALSE, symbol_txt = "*")
}
\arguments{
\item{mu_k}{- (vector) of length J_Y_z (where J_Y_z is the number of indicators)}

\item{S_k}{- (Matrix) of size J_Y_z by J_Y_z}

\item{usev_vec}{(vector) of charcters with the names of the latent class indicators}

\item{Model_Covariance}{- (logical) independence model or not. Defaults to FALSE.}

\item{symbol_txt}{("character") Either "*" or the "(at)" symbol. The latter for fixing parameters.}
}
\value{
Returns a character vector with Mplus code.
}
\description{
This function obtain imputations using Amelia, stratified by observed group membership
}
\examples{
create_withinclass_Mplus_syntax(mu_k, S_k, usev_vec)
}
