% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuke_dirs.R
\name{nuke_dirs}
\alias{nuke_dirs}
\title{Permanently remove folers and subfolders}
\usage{
nuke_dirs(directories, ...)
}
\arguments{
\item{directories}{(vector) Character strings of directories}

\item{...}{other arguments passed to the unlink() function.}
}
\description{
Permanently remove folers and subfolders
}
