% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plist2getFMM.R
\name{Plist2getFMM}
\alias{Plist2getFMM}
\title{Converts from Plist to get_FMM list needed to make input files}
\usage{
Plist2getFMM(Plist_x)
}
\arguments{
\item{Plist_x}{}
}
\description{
Converts from Plist to get_FMM list needed to make input files
}
