% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plist2Mplus.R
\name{Plist2Mplus}
\alias{Plist2Mplus}
\title{Convert population Plist to Mplus format}
\usage{
Plist2Mplus(z, data_conditions)
}
\arguments{
\item{z}{(integer)}

\item{data_conditions}{(data.frame)}
}
\value{
(data.frame) with population parameter values
}
\description{
Convert population Plist to Mplus format
}
