% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_subroutine.R
\name{pool_subroutine}
\alias{pool_subroutine}
\title{Pooling subroutine called from tracker2_pool}
\usage{
pool_subroutine(oneline_df, data_conditions, M_max, M_min = 5,
  calculate.ARIU = TRUE, calculate.KL = TRUE,
  points.montecarlo = 50000, deviance_coverage = TRUE,
  save_details = FALSE)
}
\arguments{
\item{oneline_df}{- (data.frame) containing a single line in from pick1_df in tracker2_pool for only one of the imputed files}

\item{data_conditions}{- (data.frame)}

\item{M_max}{- (integer). The number of imputations to use}

\item{M_min}{- (integer). Optional. Defaults to a requirement of at least 5 imputations to report results.}

\item{calculate.ARIU}{(logical) Defaults to TRUE. Whether to report the average relative increase in variance with reference to the complete data set}

\item{calculate.KL}{(logical) Defaults to TRUE. Whether to calculate an estimate of the KL divergence between the model implied by the the pooled point estimates and the population model.}

\item{points.montecarlo}{(integer) Defaults to 5E4}

\item{deviance_coverage}{(logical) Defaults to TRUE. Will calculate deviance and coverage of each parameter from population values. Only available if kfit == 3}

\item{save_details}{(logical) Defaults to FALSE. If true, then also save the individual point estimates, within imputation asymptotic cov matrix, and between imputation estimate}
}
\value{
(data.frame) with updated line in tracker, including location of saved file analyzing results
}
\description{
Pooling subroutine called from tracker2_pool
}
