% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlist2Mplus.R
\name{Qlist2Mplus}
\alias{Qlist2Mplus}
\title{Convert estimated Qlist to Mplus format}
\usage{
Qlist2Mplus(Qlist, z, data_conditions)
}
\arguments{
\item{Qlist}{(list)}

\item{z}{(integer)}

\item{data_conditions}{(data.frame)}
}
\value{
(data.frame) with estimates in Mplus parameter format
}
\description{
Convert estimated Qlist to Mplus format
}
