% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlist2syntax.R
\name{Qlist2syntax}
\alias{Qlist2syntax}
\title{Qlist2syntax.R}
\usage{
Qlist2syntax(Qlist, symbol = "@")
}
\arguments{
\item{symbol}{- (character) Defaults to "(at)"}

\item{Qlist-}{(list)}
}
\value{
(character vector)
}
\description{
Qlist to svals syntax
}
