% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_label_switch.R
\name{resolve_label_switch}
\alias{resolve_label_switch}
\title{Resolve label switching with population values known.}
\usage{
resolve_label_switch(out_ftc, name = "glpk", t_max = 30,
  approximate = 0, round_cplex = 0, trace_cplex = 0, ...)
}
\arguments{
\item{out_ftc}{(list) An object from fit_til_convergence.}

\item{name}{(character) Defaults to "glpk"}

\item{t_max}{(integer) Maximum time for estimation using "glpk". Defaults to 30 sec.}

\item{approximate}{(integer) Allow for approximate solution. Defauts to 0 (i.e., FALSE)}

\item{round_cplex}{(integer)}

\item{trace_cplex}{(integer)}

\item{...}{Other arguments to send to create_naiveMplus_inpfile()}
}
\value{
(list) A list with the following things:
          (a) class_switched - (logical) identifies if class switching had occurred.
          (b) map_from - (vector) always just 1:K showing class definitions from sample
          (c) map_to - (NULL or vector) same as map_from if no class switching.
                     Otherwise it is the corresponding population-based classes from sample classes
          (d) out_Mplus- (list) readModels() fitting model with starting values that resolve class switching
}
\description{
Resolve label switching with population values known.
}
\examples{




}
