% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_pool.R
\name{results_pool}
\alias{results_pool}
\title{Get results}
\usage{
results_pool(p, temp_wd_p_vec, kk, z, rep, starts_txt)
}
\arguments{
\item{p}{(integer) Processor number}

\item{kk}{(integer) Fitted number of classes}

\item{z}{}
}
\value{
(data.frame) results_df
}
\description{
Get results
}
\examples{



}
