% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixmvrnorm.R
\name{rmixmvrnorm}
\alias{rmixmvrnorm}
\title{Samples from mixture of MVNs}
\usage{
rmixmvrnorm(N, pi_vec, mu_mat, S_array)
}
\arguments{
\item{N}{(integer) Number of samples to draw}

\item{pi_vec}{(numeric vector) with K elements containing component probabilities}

\item{mu_mat}{(matrix) of size JxK, where J is number of items}

\item{S_array}{(array) of size JxJxK with within class covariance matrix.}
}
\value{
(matrix) of size NxJ  containing draws from finite ixture of MVNs
}
\description{
This function draws samples from a finite mixture of MVNs
}
\examples{
rmixmvrnorm(N, pi_vec, mu_mat, S_array)
}
