% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratamelia.R
\name{stratamelia}
\alias{stratamelia}
\title{Stratified Amelia Imputation}
\usage{
stratamelia(x, strata, m, ...)
}
\arguments{
\item{x}{(data.frame) Incomplete data set}

\item{strata}{(vector of integers) Group membership indicator}

\item{...}{additional arguments passed to the amelia function}
}
\value{
out_list (list) with each element being an imputed data set.
}
\description{
This function obtain imputations using Amelia, stratified by observed group membership
}
\examples{
stratamelia(x, strata, m = 1)
}
