% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_and_get.R
\name{switch_and_get}
\alias{switch_and_get}
\title{Resolve label switching and extracts parameter estimates.}
\usage{
switch_and_get(out_ftc, z, data_conditions)
}
\arguments{
\item{out_ftc}{(list) An object from fit_til_convergence. Currently the options readModels = TRUE and savedata = TRUE must be specified in fit_til_convergence.}

\item{z}{(integer) Data conditioning number}

\item{data_conditions}{(data.frame)}
}
\value{
(data.frame) with parameter estimates
}
\description{
Resolve label switching and extracts parameter estimates.
}
\examples{
param_complete = switch_and_get(out_ftc = out_ftc_complete, z = z, data_conditions = data_conditions)
}
