% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax2Plist.R
\name{syntax2Plist}
\alias{syntax2Plist}
\title{Convert Mplus MODEL: syntax to a Plist}
\usage{
syntax2Plist(syntax_x, z_x, data_conditions_x, Plist_x = NULL,
  out_get_FMM_x = NULL, C_letter = "C", Y_letter = "Y")
}
\arguments{
\item{syntax_x}{(character)}

\item{z_x}{(integer)}

\item{data_conditions_x}{(data.frame)}

\item{Plist_x}{(list) Optional. Defaults to NULL. If not specified, then extracts using get_Plist}

\item{out_get_FMM_x}{(list)}

\item{C_letter}{(character) Optional. Defaults to "C".}

\item{Y_letter}{(character) Optional. Defaults to "Y"}
}
\value{
(list)
}
\description{
Convert Mplus MODEL: syntax to a Plist
}
