% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax2Qlist.R
\name{syntax2Qlist}
\alias{syntax2Qlist}
\title{Convert Mplus MODEL: syntax to a Qlist}
\usage{
syntax2Qlist(syntax_x, Plist_x, C_letter = "C", Y_letter = "Y",
  fix_S = TRUE)
}
\arguments{
\item{syntax_x}{(character)}

\item{Plist_x}{(list) Optional. Defaults to NULL. If not specified, then extracts using get_Plist}

\item{C_letter}{(character) Optional. Defaults to "C".}

\item{Y_letter}{(character) Optional. Defaults to "Y"}

\item{fix_S}{(logical) Defaults to true.}
}
\value{
(list)
}
\description{
Convert Mplus MODEL: syntax to a Qlist
}
