% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_outlist.R
\name{to_outlist}
\alias{to_outlist}
\title{Compile fitted models to an outlist for complete and observed data only}
\usage{
to_outlist(out_ftc, out_switched, Data, p, z, rep, pm = NA, pva = NA)
}
\arguments{
\item{out_ftc}{(list) from out_ftc()}

\item{out_switched}{(list) from resolve_label_switch}

\item{Data}{(character) indicating}

\item{p}{(integer)}

\item{z}{(integer)}

\item{rep}{(integer)}

\item{pm}{(integer) defaults to NA}

\item{pva}{(integer) defaults to NA}
}
\description{
Compile fitted models to an outlist for complete and observed data only
}
\examples{

}
