% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2_fit_til_convergence.R
\name{tracker2_fit_til_convergence}
\alias{tracker2_fit_til_convergence}
\title{Fit to convergence using a tracker}
\usage{
tracker2_fit_til_convergence(data_conditions, tracker_df, itermax = 2,
  multiplier = 2^5, Rcond_min = 1e-06, core_limit = NULL,
  cl = NULL)
}
\arguments{
\item{data_conditions}{(data.frame)}

\item{tracker_df}{(data.frame)}

\item{itermax}{(numeric) Default is 4^3 (or 64). Specifying the number of minutes a model should take to fit (on average) before terminating. Defaults to 1 minute.}

\item{multiplier}{(numeric) Multiplier across each iteration by which random starts should increase}

\item{Rcond_min}{(numeric) Threshold below which results are flagged for poor conditioning number.}

\item{core_limit}{(integer) Defaults to detect.cores()-2; Number of cores to limit in fitting mixture models in mplus specifiying Processors option.}

\item{cl}{cluster object from makeCluster}

\item{Processors}{(integer) Defaults to parallel::detectCores()-1}
}
\value{
tracker_df
}
\description{
Fit to convergence using a tracker
}
\examples{
tracker2_fit_til_convergence(data_conditions, tracker_df)

}
