% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2_pool.R
\name{tracker2_pool}
\alias{tracker2_pool}
\title{Pool analysis files given tracker information}
\usage{
tracker2_pool(tracker_df, pool_wd, data_conditions, cl = NULL,
  M_max = Inf, ...)
}
\arguments{
\item{tracker_df}{(data.frame) the tracker}

\item{pool_wd}{- (character) Where the Rdata file with the pooled estimates and statistics should be saved}

\item{data_conditions}{(data.frame) the data conditoins}

\item{cl}{(cluster) A cluster variable from the SNOW package.}

\item{M_max}{- (integer) Maximum number of imputations to use}

\item{...}{Optional argument to send to pool_subroutine (e.g., M_max, M_min, calculate.ARIU, and/or calculate.KL)}
}
\value{
tracker_df - updated with saved pooled information
}
\description{
Pool analysis files given tracker information
}
