% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2_resolve_label_switch.R
\name{tracker2_resolve_label_switch}
\alias{tracker2_resolve_label_switch}
\title{Resolve label switching with population values known for a given file (row) value in a tracker.}
\usage{
tracker2_resolve_label_switch(oneline_df, Plist, out_get_FMM = NULL,
  args_KL_check_switching = list(NULL),
  args_tracker2inpfile_naive = list(NULL), showOutput = FALSE)
}
\arguments{
\item{oneline_df}{(data.frame) The tracker}

\item{Plist}{(list) List with population values to compare against}

\item{out_get_FMM}{Optional. Default to NULL. If not provided, calculated using Plist with the Plist2getFMM() function.}

\item{args_KL_check_switching}{(list) Optional arguments pasted to KL_check_switching() beyond nn, mu_est_mat, S_est_array, z, data_conditoins}

\item{args_tracker2inpfile_naive}{(list) Optional arguments passed to tracker2inpfile_naive() beyond nn, tracker_df, pop_params_nn, data_conditions, Model_txt, process_cores}

\item{showOutput}{(logical) Defaluts to FALSE. This sets the showOutput argument in MplusAutomation::runModels()}
}
\value{
Refits and generates new .out file correcting for switching. Returns a data.frame.
}
\description{
Resolve label switching with population values known for a given file (row) value in a tracker.
}
\examples{
tracker2_resolve_label_switch<-function(nn,tracker_df, data_conditions,
}
