% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2_save_readMplus.R
\name{tracker2_save_readMplus}
\alias{tracker2_save_readMplus}
\title{Reads in .out files and saves the results}
\usage{
tracker2_save_readMplus(tracker_df, estwd, cl = NULL,
  args_readModels = NULL)
}
\arguments{
\item{tracker_df}{(data.frame) the tracker}

\item{estwd}{(character) working directory for the estimates}

\item{cl}{Optional cluster object .}

\item{args_readModels}{(list) Additional arguments to MplusAutomation::readModels()}
}
\value{
tracker_df
}
\description{
Reads in .out files and saves the results
}
