% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2_switch.R
\name{tracker2_switch}
\alias{tracker2_switch}
\title{Address label switching from a tracker of .out files}
\usage{
tracker2_switch(tracker_df, data_conditions, cl = NULL,
  args_KL_check_switching = NULL, args_tracker2inpfile_naive = NULL)
}
\arguments{
\item{tracker_df}{(data.frame) The tracker}

\item{data_conditions}{(data.frame) data conditions}

\item{cl}{optional cluster object}

\item{args_KL_check_switching}{(list) Additional arguments to lpa.mi.src::KL_check_switching}

\item{args_tracker2inpfile_naive}{(list) additional argument to lpa.mi.src::tracker2inpfile_naive()}
}
\description{
Address label switching from a tracker of .out files
}
