% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2fitfixed.R
\name{tracker2fitfixed}
\alias{tracker2fitfixed}
\title{tracker2fitfixed}
\usage{
tracker2fitfixed(oneline_df, args_tracker2inpfile_naive = list(NULL),
  showOutput = FALSE)
}
\arguments{
\item{oneline_df}{(data.frame) One row from the tracker data frame for monitoring covernce, specifying dat file, specifying .inp file path, etc.}

\item{args_tracker2inpfile_naive}{(list) Optional arguments passed to tracker2inpfile_naive() beyond nn, tracker_df, pop_params_nn, data_conditions, Model_txt, process_cores}

\item{showOutput}{(logical) Defaluts to FALSE. This sets the showOutput argument in MplusAutomation::runModels()}
}
\value{
Saves the Mplus input files for Naive LPA analysis in the the folders specified of dffolderfiles input. Returns a character vector with the the Mplus code.
}
\description{
Creates an input file that fits an LPA model fixed at the pooled parameter estiamtes given information from a tracker data frame.
}
\examples{
tracker2fitfixed(oneline_df)
}
