% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2inpfile_naive.R
\name{tracker2inpfile_naive}
\alias{tracker2inpfile_naive}
\title{Create Mplus .inp file from a tracker}
\usage{
tracker2inpfile_naive(oneline_df, pop_params_nn = NULL,
  data_conditions = NULL, savedata = FALSE, estimates = FALSE,
  results = FALSE, save_tech3 = FALSE, symbol_txt = "*",
  Model_txt = NULL, weight = FALSE, output_txt = NULL,
  processor_cores = 1, do_dk = FALSE)
}
\arguments{
\item{oneline_df}{(data.frame) One row from the tracker data frame for monitoring covernce, specifying dat file, specifying .inp file path, etc.}

\item{pop_params_nn}{(list) Optional. Gaussian finite mixture model parameters in format exported by get_FMM_params for the nn-th condition in the tracker. Will obtain if not specified.}

\item{data_conditions}{(data.frame) Optional, but must be provided if pop_params_nn not given.}

\item{savedata}{(logical) Save the cpros or not (savedata = FALSE). Defaults to false.}

\item{estimates}{(logical) Save the parameter estimates (i.e. include a ESTIMATES = ... statement in the Savedata Mplus command). Defaults to FALSE.}

\item{results}{(logical) Save the results (i.e. include a RESULTS = ... statement in the Savedata Mplus command). Defaults to FALSE.}

\item{save_tech3}{(logical) Save the asympototic variance covariance matrix. Defaults to FALSE.}

\item{symbol_txt}{("character") Either "*" or the "(at)" symbol. The latter for fixing parameters. Defaults to star.}

\item{Model_txt}{(character vector). Syntax for inside MODEL: command (exclude a "MODEL:" line when specifying). Defaults to NULL.}

\item{weight}{(logical) Indicates if a sample weight is included as the last variable in the .dat files listed in dffolderfiles. Defaults to FALSE.}

\item{output_txt}{(character) Optional. Text for Output block to be requested in addition to the default 'tech1;', 'svalues;', 'tech3;'}

\item{processor_cores}{(integer) Number of processor cores to use. Defaults to 1.}

\item{do_dk}{(logical) Tells me whether or not fit a model with smaller number of components}
}
\value{
Saves the Mplus input files for Naive LPA analysis in the the folders specified of dffolderfiles input. Returns a character vector with the the Mplus code.
}
\description{
Creates an input file that fits a simple (i.e., naive) LPA model given information from a tracker data frame.
}
\examples{
tracker2inpfile_naive(oneline_df)
}
