% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker2resolve_label_switch.R
\name{tracker2resolve_label_switch}
\alias{tracker2resolve_label_switch}
\title{Resolve label switching with population values known.}
\usage{
tracker2resolve_label_switch(nn, tracker_df, data_conditions,
  out_get_FMM = NULL, args_KL_check_switching = list(NULL),
  args_tracker2inpfile_naive = list(NULL))
}
\arguments{
\item{nn}{(integer) Identifier in tracker_df}

\item{tracker_df}{(data.frame) The tracker}

\item{args_KL_check_switching}{(list) Optional arguments pasted to KL_check_switching()}

\item{args_tracker2inpfile_naive}{(list) Optional arguments passed to tracker2inpfile_naive()}

\item{get_FMM_params}{(list) Optional. From get_FMM_params}
}
\value{
Refits and generates new .out file correcting for switching. Returns a data.frame.
}
\description{
Resolve label switching with population values known.
}
\examples{
tracker2resolve_label_switch<-function(nn,tracker_df, data_conditions,
}
