% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence}
\usage{
check_convergence(file, folder_wd = get.wd(), starts_txt = "0;")
}
\arguments{
\item{file}{(integer)}
}
\value{
(logical)
}
\description{
Check convergence
}
