% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_naiveMplus_inpfile.R
\name{create_naiveMplus_inpfile}
\alias{create_naiveMplus_inpfile}
\title{Create Mplus .inp file}
\usage{
create_naiveMplus_inpfile(z, out_get_FMM, dffolderfiles, temp_wd_p,
  savedata = FALSE, estimates = FALSE, results = FALSE,
  save_tech3 = FALSE, type_imputation = FALSE, starts_txt = "0;",
  symbol_txt = "*", Model_txt = NULL, weight = FALSE)
}
\arguments{
\item{z}{(integer) Condition number}

\item{out_get_FMM}{(list) Gaussian finite mixture model parameters in format exported by get_FMM_params}

\item{dffolderfiles}{(data.frame) with the files and folders of the saved data}

\item{temp_wd_p}{(character). Specifies processor-specific temporary directory}

\item{savedata}{(logical) Save the cpros or not (savedata = FALSE). Defaults to false.}

\item{estimates}{(logical) Save the parameter estimates (i.e. include a ESTIMATES = ... statement in the Savedata Mplus command). Defaults to FALSE.}

\item{results}{(logical) Save the results (i.e. include a RESULTS = ... statement in the Savedata Mplus command). Defaults to FALSE.}

\item{save_tech3}{(logical) Save the asympototic variance covariance matrix. Defaults to FALSE.}

\item{type_imputation}{(logical) Add a "TYPE = IMPUTATION;" command in the Data command of Mplus. Defaults to FALSE (i.e. not TYPE statement in DATA command)}

\item{starts_txt}{(character) Custom starting values in Mplus (e.g "20 10;" correpsonds to "STARTS = 20 10;"). Defaults to "0;"}

\item{symbol_txt}{("character") Either "*" or the "(at)" symbol. The latter for fixing parameters.}

\item{Model_txt}{(character vector). Defaults to NULL.}

\item{weight}{(logical) Indicates if a sample weight is included as the last variable in the .dat files listed in dffolderfiles.}
}
\value{
Saves the Mplus input files for Naive LPA analysis in the the folders specified of dffolderfiles input. Returns a character vector with the the Mplus code.
}
\description{
Creates an input file that fits a simple (i.e., naive) LPA model
}
\examples{
create_naiveMplus_inpfile(z, data_conditions, dffolderfiles, temp_wd_p)
}
