% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ICs_naive.R
\name{extract_ICs_naive}
\alias{extract_ICs_naive}
\title{Get ICs}
\usage{
extract_ICs_naive(is_Complete, p, temp_wd_p_vec, kk, z, rep, starts_txt,
  pm = NA, pva = NA)
}
\arguments{
\item{is_Complete}{(logical) If true, then looks in ~/Complete data folder; otherwise, reads out of ~/Observed data}

\item{p}{(integer) Processor number}

\item{kk}{(integer) Fitted number of classes}

\item{pm}{Defaults to NA}

\item{pva}{Defaults to NA}
}
\value{
(data.frame) results_df
}
\description{
Get ICs
}
\examples{



}
