% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_do_rubinsrules.R
\name{fit_and_do_rubinrules}
\alias{fit_and_do_rubinrules}
\title{Fit mixture model to multiply imputed data and pool using Rubin's Rules (1987)}
\usage{
fit_and_do_rubinrules(dff_imputed, methods_list, pop_params_z,
  data_conditions, temp_wd_p_vec, p, z, pm, pva)
}
\arguments{
\item{dff_imputed}{(data.frame)}

\item{methods_list}{(list)}

\item{pop_params_z}{(list)}

\item{data_conditions}{(data.frame)}

\item{temp_wd_p_vec}{(vector) of characters}

\item{p}{(integer)}

\item{z}{(integer)}

\item{pm}{(integer)}

\item{pva}{(integer)}
}
\value{
(list) out_list with:
              (a) problem - (logical) indicating if less than minimum imputations converged
              (b) MM_pva - (integer) Number of imputation draws that converged
              (c) parameters - (data.frame) Pooled parameters in MplusAutomation format
              (d) Vtot - (matrix) Pooled variance taking into account within- (Vbar) & between- (B) imputation variances
              (e) Vbar - (matrix) Within imptuation variance
              (f) B - (matrix) Between imputation variance V_m_list,
              (g) theta_m_list - (list) with individual imputation point estimates (as a vector)
}
\description{
Fit mixture model to multiply imputed data and pool using Rubin's Rules (1987)
}
\examples{
fit_and_do_rubinrules(dff_imputed, methods_list, pop_params_z, data_conditions, temp_wd_p_vec, p, z, pm, pva)

}
